restore hasContainers & apply the logic to delete of a resource (wip)

+++ src/Models/AbstractEntity.php
@@ -82,6 +82,7 @@ use PDOStatement;
 use Override;
 use Symfony\Component\HttpFoundation\InputBag;
 use Symfony\Component\HttpFoundation\Request;
+use Throwable;
 use ZipArchive;
 
 use function array_column;
@@ -846,18 +847,28 @@ abstract class AbstractEntity extends AbstractRest
     }
 
     #[Override]
-    public function destroy(): bool
+    public function destroy(bool $deleteContainers = false): bool
     {
         $this->canOrExplode(AccessType::Write);
-        // remove the custom_id upon deletion
-        $this->update(new EntityParams('custom_id', ''));
-        // delete from pinned too
-        new Pins($this)->cleanup();
-        $this->Uploads->destroyAll();
-        // entity deletion is soft, but its container rows must be removed permanently.
-        // Otherwise, invisible containers keep their storage locations occupied (see #6418).
-        LinksFactory::getContainersLinks($this)->destroyAll();
-        return $this->update(new EntityParams('state', State::Deleted->value));
+        $this->Db->beginTransaction();
+        try {
+            // remove the custom_id upon deletion
+            $this->update(new EntityParams('custom_id', ''));
+            // delete from pinned too
+            new Pins($this)->cleanup();
+            $this->Uploads->destroyAll();
+            if ($deleteContainers) {
+                LinksFactory::getContainersLinks($this)->destroyAll();
+            }
+            $result = $this->update(new EntityParams('state', State::Deleted->value));
+            if ($deleteContainers) {
+                $this->Db->commit();
+            }
+            return $result;
+        } catch (Throwable $e) {
+            $this->Db->rollBack();
+            throw $e;
+        }
     }
 
     public function restore(): bool
